+++ examples/data-sources/nxos_ipv4_access_list/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_ipv4_access_list" "example" {
-  name = "ACL1"
-}