+++ examples/data-sources/nxos_ipv4_access_list_entry/data-source.tf
@@ -1,4 +1,0 @@
-data "nxos_ipv4_access_list_entry" "example" {
-  name            = "ACL1"
-  sequence_number = 10
-}