+++ examples/data-sources/nxos_ipv4_access_list_policy_ingress_interface/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_ipv4_access_list_policy_ingress_interface" "example" {
-  interface_id = "eth1/10"
-}