+++ examples/data-sources/nxos_ipv4_access_lists/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_ipv4_access_lists" "example" {
+}