+++ examples/resources/nxos_ipv4_access_list/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_ipv4_access_list.example
-  identity = {
-    "name" : "<name>"
-  }
-}