+++ examples/resources/nxos_ipv4_access_list/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_ipv4_access_list.example
-  id = "<name>"
-}