+++ examples/resources/nxos_ipv4_access_list/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv4_access_list.example "<name>"