+++ examples/resources/nxos_ipv4_access_list_entry/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv4_access_list_entry.example "sys/acl/ipv4/name-[ACL1]/seq-[10]"