+++ examples/resources/nxos_ipv4_access_list_policy_egress_interface/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_ipv4_access_list_policy_egress_interface.example
-  identity = {
-    "interface_id" : "<interface_id>"
-  }
-}