+++ examples/resources/nxos_ipv4_access_list_policy_egress_interface/import-by-string-id.tf
@@ -1,4 +1,0 @@
-import {
-  to = nxos_ipv4_access_list_policy_egress_interface.example
-  id = "<interface_id>"
-}