+++ examples/resources/nxos_ipv4_access_list_policy_ingress_interface/import.sh
@@ -1,1 +1,0 @@
-terraform import nxos_ipv4_access_list_policy_ingress_interface.example "<interface_id>"