+++ examples/resources/nxos_ipv4_access_list_policy_ingress_interface/resource.tf
@@ -1,4 +1,0 @@
-resource "nxos_ipv4_access_list_policy_ingress_interface" "example" {
-  interface_id     = "eth1/10"
-  access_list_name = "ACL1"
-}