+++ examples/resources/nxos_ipv4_access_lists/import-by-identity.tf
@@ -1,0 +1,5 @@
+import {
+  to = nxos_ipv4_access_lists.example
+  identity = {
+  }
+}