+++ examples/resources/nxos_ipv4_access_lists/import-by-string-id.tf
@@ -1,0 +1,4 @@
+import {
+  to = nxos_ipv4_access_lists.example
+  id = ""
+}