+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - BREAKING CHANGE: Remove `nxos_ntp_server` resource and data source, NTP configuration is now managed as a single `nxos_ntp` resource and data source
 - BREAKING CHANGE: Remove `nxos_dhcp_relay_interface` resource and data source and `nxos_dhcp_relay_address` resource and data source, DHCP configuration is now managed as a single `nxos_dhcp` resource and data source
 - BREAKING CHANGE: Remove `nxos_spanning_tree_interface` resource and data source, Spanning Tree configuration is now managed as a single `nxos_spanning_tree` resource and data source
+- BREAKING CHANGE: Remove `nxos_ipv4_access_list` resource and data source, `nxos_ipv4_access_list_policy_egress_interface` resource and data source, and `nxos_ipv4_access_list_policy_ingress_interface` resource and data source, IPv4 Access List configuration is now managed as a single `nxos_ipv4_access_lists` resource and data source
 
 ## 0.7.0